Adding description and value attributes to ProductCard

+++ app/src/main/java/com/ronellyson/smart_fast_food/data/dto/ProductDTO.java
@@ -3,6 +3,7 @@ package com.ronellyson.smart_fast_food.data.dto;
 import com.ronellyson.smart_fast_food.data.model.Product;
 import com.ronellyson.smart_fast_food.data.network.response.ProductResponse;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -11,7 +12,7 @@ public class ProductDTO {
         List<Product> foods = new ArrayList<>();
 
         for(ProductResponse response: productResponse){
-            Product food = new Product(response.getName(), response.getImage(), response.getPrice(),
+            Product food = new Product(response.getName(), response.getImage(), BigDecimal.valueOf(response.getPrice()),
                     response.getDescription(), response.getRate());
 
             foods.add(food);
